requirements.txt
Pillow==8.1.0
Markdown==3.3.4
numpy==1.19.5
plotly==5.5.0
pandas==1.2.0
streamlit==1.4.0
openpyxl==3.0.6
matplotlib==3.4.2
matplotlib-inline==0.1.2
openpyxl==3.0.6
pyxlsb==1.0.9
XlsxWriter==3.0.2